chore: update config structure
