+++ src/lib/config/index.ts
@@ -1,0 +1,2 @@
+export * from '@/lib/config/site-config';
+export * from '@/lib/config/locomotive-config';
