+++ src/lib/config/locomotive-config.ts
@@ -1,29 +1,3 @@
-export const siteConfig = {
-  name: 'Derail Valley Planner',
-  description: 'A route planning tool for Derail Valley',
-  keywords: ['derail valley', 'route planner', 'train simulator'],
-  authors: {
-    devTotoro: { name: 'DevTotoro', url: 'https://github.com/DevTotoro' }
-  },
-  links: {
-    github: 'https://github.com/DevTotoro/derail-valley-planner'
-  },
-  navItems: [
-    {
-      label: 'Deliveries',
-      href: '/deliveries'
-    },
-    {
-      label: 'Locomotives',
-      href: '/locomotives'
-    },
-    {
-      label: 'Stations',
-      href: '/stations'
-    }
-  ]
-};
-
 export const locomotiveTypes = {
   dm: 'Diesel-Mechanical',
   de: 'Diesel-Electric',
