+++ src/lib/config/site-config.ts
@@ -1,0 +1,25 @@
+export const siteConfig = {
+  name: 'Derail Valley Planner',
+  description: 'A route planning tool for Derail Valley',
+  keywords: ['derail valley', 'route planner', 'train simulator'],
+  authors: {
+    devTotoro: { name: 'DevTotoro', url: 'https://github.com/DevTotoro' }
+  },
+  links: {
+    github: 'https://github.com/DevTotoro/derail-valley-planner'
+  },
+  navItems: [
+    {
+      label: 'Deliveries',
+      href: '/deliveries'
+    },
+    {
+      label: 'Locomotives',
+      href: '/locomotives'
+    },
+    {
+      label: 'Stations',
+      href: '/stations'
+    }
+  ]
+};
